+++ src/TrendyolClient.Sharp/Models/TrendyolRequestAlternativeDelivery.cs
@@ -18,5 +18,5 @@ public class TrendyolRequestAlternativeDelivery
     public int? BoxQuantity { get; set; }
 
     [JsonPropertyName("deci")]
-    public double? Deci { get; set; }
+    public decimal? Deci { get; set; }
 }
