+++ src/TrendyolClient.Sharp/Models/TrendyolSettlementTransactionData.cs
@@ -48,13 +48,13 @@ public class TrendyolSettlementTransactionData
     ///     Debt amount
     /// </summary>
     [JsonPropertyName("debt")]
-    public double Debt { get; set; }
+    public decimal Debt { get; set; }
 
     /// <summary>
     ///     Credit amount
     /// </summary>
     [JsonPropertyName("credit")]
-    public double Credit { get; set; }
+    public decimal Credit { get; set; }
 
     /// <summary>
     ///     Payment period in days
@@ -66,19 +66,19 @@ public class TrendyolSettlementTransactionData
     ///     Commission rate percentage
     /// </summary>
     [JsonPropertyName("commissionRate")]
-    public double? CommissionRate { get; set; }
+    public decimal? CommissionRate { get; set; }
 
     /// <summary>
     ///     Commission amount
     /// </summary>
     [JsonPropertyName("commissionAmount")]
-    public double? CommissionAmount { get; set; }
+    public decimal? CommissionAmount { get; set; }
 
     /// <summary>
     ///     Seller revenue amount
     /// </summary>
     [JsonPropertyName("sellerRevenue")]
-    public double? SellerRevenue { get; set; }
+    public decimal? SellerRevenue { get; set; }
 
     /// <summary>
     ///     Order number
